refactor: resolve #870 - split sampleBgData.ts (711 lines) exceeding 500-line limit
Closes #870
