+++ src/core/samples/sampleBgData.ts
@@ -1,649 +1,35 @@
-/* eslint-disable max-len */
-/* eslint-disable @typescript-eslint/naming-convention */
+ 
 /**
  * Sample BG Data for programs that use the VIEW command
  *
  * These are pre-designed BG graphics that pair with sample programs.
  * Each BG is a 28x21 grid of cells with character codes and color patterns.
- */
-
-import { BG_GRID, DEFAULT_BG_CHAR_CODE } from '@/features/bg-editor/constants'
-import type { BgCell, BgGridData } from '@/features/bg-editor/types'
-
-/** Create an empty 28x21 grid */
-function createEmptyGrid(): BgGridData {
-  const grid: BgGridData = []
-  for (let y = 0; y < BG_GRID.ROWS; y++) {
-    const row: BgCell[] = []
-    for (let x = 0; x < BG_GRID.COLS; x++) {
-      row.push({ charCode: DEFAULT_BG_CHAR_CODE, colorPattern: 0 })
-    }
-    grid.push(row)
-  }
-  return grid
-}
-
-/** Helper to set a cell in the grid */
-function setCell(grid: BgGridData, x: number, y: number, charCode: number, colorPattern: 0 | 1 | 2 | 3 = 0): void {
-  if (x >= 0 && x < BG_GRID.COLS && y >= 0 && y < BG_GRID.ROWS) {
-    const row = grid[y]
-    if (row) {
-      row[x] = { charCode, colorPattern }
-    }
-  }
-}
-
-/**
- * Demo BG with a simple border and centered text area
- * Shows the grid boundaries clearly
- */
-function createDemoBorderGrid(): BgGridData {
-  const grid = createEmptyGrid()
-
-  // Draw border using block characters (176-183 are BG items)
-  // Top and bottom borders
-  for (let x = 0; x < BG_GRID.COLS; x++) {
-    setCell(grid, x, 0, 176, 1) // Top border
-    setCell(grid, x, BG_GRID.ROWS - 1, 176, 1) // Bottom border
-  }
-
-  // Left and right borders
-  for (let y = 0; y < BG_GRID.ROWS; y++) {
-    setCell(grid, 0, y, 177, 1) // Left border
-    setCell(grid, BG_GRID.COLS - 1, y, 177, 1) // Right border
-  }
-
-  // Corner pieces
-  setCell(grid, 0, 0, 178, 2)
-  setCell(grid, BG_GRID.COLS - 1, 0, 178, 2)
-  setCell(grid, 0, BG_GRID.ROWS - 1, 178, 2)
-  setCell(grid, BG_GRID.COLS - 1, BG_GRID.ROWS - 1, 178, 2)
-
-  // Center decoration - simple pattern
-  const centerX = Math.floor(BG_GRID.COLS / 2)
-  const centerY = Math.floor(BG_GRID.ROWS / 2)
-
-  // Draw a diamond shape in center
-  setCell(grid, centerX, centerY - 2, 42, 2) // *
-  setCell(grid, centerX - 1, centerY - 1, 42, 2)
-  setCell(grid, centerX, centerY - 1, 42, 2)
-  setCell(grid, centerX + 1, centerY - 1, 42, 2)
-  setCell(grid, centerX - 2, centerY, 42, 2)
-  setCell(grid, centerX - 1, centerY, 65, 3) // A
-  setCell(grid, centerX, centerY, 42, 2)
-  setCell(grid, centerX + 1, centerY, 42, 2)
-  setCell(grid, centerX + 2, centerY, 42, 2)
-  setCell(grid, centerX - 1, centerY + 1, 42, 2)
-  setCell(grid, centerX, centerY + 1, 42, 2)
-  setCell(grid, centerX + 1, centerY + 1, 42, 2)
-  setCell(grid, centerX, centerY + 2, 42, 2)
-
-  return grid
-}
-
-/**
- * Platform game level background
- * Has ground, platforms, and decorative elements
- */
-function createPlatformLevelGrid(): BgGridData {
-  const grid = createEmptyGrid()
-
-  // Ground (bottom 2 rows) - solid blocks
-  for (let x = 0; x < BG_GRID.COLS; x++) {
-    setCell(grid, x, BG_GRID.ROWS - 1, 219, 1) // Solid block
-    setCell(grid, x, BG_GRID.ROWS - 2, 220, 1) // Half block
-  }
-
-  // Platforms at different heights
-  // Platform 1: left side, mid height
-  for (let x = 2; x < 8; x++) {
-    setCell(grid, x, 14, 223, 2) // Top half block
-    setCell(grid, x, 15, 219, 2) // Solid block
-  }
-
-  // Platform 2: center, higher
-  for (let x = 10; x < 18; x++) {
-    setCell(grid, x, 10, 223, 2)
-    setCell(grid, x, 11, 219, 2)
-  }
-
-  // Platform 3: right side, mid height
-  for (let x = 20; x < 26; x++) {
-    setCell(grid, x, 14, 223, 2)
-    setCell(grid, x, 15, 219, 2)
-  }
-
-  // Floating blocks (coin-like)
-  setCell(grid, 5, 8, 36, 3) // $ - coin
-  setCell(grid, 14, 6, 36, 3)
-  setCell(grid, 22, 8, 36, 3)
-
-  // Clouds (using * and @ characters)
-  setCell(grid, 4, 3, 42, 0) // *
-  setCell(grid, 5, 3, 42, 0)
-  setCell(grid, 6, 3, 42, 0)
-  setCell(grid, 5, 2, 42, 0)
-
-  setCell(grid, 18, 4, 42, 0)
-  setCell(grid, 19, 4, 42, 0)
-  setCell(grid, 20, 4, 42, 0)
-  setCell(grid, 19, 3, 42, 0)
-
-  // Pipes (using | and _)
-  setCell(grid, 24, 17, 186, 1) // Vertical line
-  setCell(grid, 24, 18, 186, 1)
-  setCell(grid, 24, 19, 186, 1)
-  setCell(grid, 23, 17, 205, 1) // Horizontal line (top)
-  setCell(grid, 25, 17, 205, 1)
-
-  return grid
-}
-
-/**
- * Title screen background
- * Decorative border with space for title text
- */
-function createTitleScreenGrid(): BgGridData {
-  const grid = createEmptyGrid()
-
-  // Double-line border using box drawing characters
-  // Top border
-  for (let x = 1; x < BG_GRID.COLS - 1; x++) {
-    setCell(grid, x, 1, 205, 2)
-    setCell(grid, x, 2, 196, 1)
-  }
-
-  // Bottom border
-  for (let x = 1; x < BG_GRID.COLS - 1; x++) {
-    setCell(grid, x, BG_GRID.ROWS - 2, 205, 2)
-    setCell(grid, x, BG_GRID.ROWS - 3, 196, 1)
-  }
-
-  // Side borders
-  for (let y = 2; y < BG_GRID.ROWS - 2; y++) {
-    setCell(grid, 1, y, 186, 2)
-    setCell(grid, 2, y, 179, 1)
-    setCell(grid, BG_GRID.COLS - 2, y, 186, 2)
-    setCell(grid, BG_GRID.COLS - 3, y, 179, 1)
-  }
-
-  // Corner pieces
-  setCell(grid, 1, 1, 201, 2) // Top-left
-  setCell(grid, BG_GRID.COLS - 2, 1, 187, 2) // Top-right
-  setCell(grid, 1, BG_GRID.ROWS - 2, 200, 2) // Bottom-left
-  setCell(grid, BG_GRID.COLS - 2, BG_GRID.ROWS - 2, 188, 2) // Bottom-right
-
-  // Stars scattered around
-  const starPositions: Array<[number, number]> = [
-    [4, 4], [8, 5], [20, 4], [24, 6],
-    [5, 16], [22, 17], [10, 18], [18, 16],
-  ]
-  for (const [x, y] of starPositions) {
-    setCell(grid, x, y, 42, 3) // *
-  }
-
-  // Decorative hearts
-  const heartPositions: Array<[number, number]> = [
-    [6, 8], [22, 8], [6, 14], [22, 14],
-  ]
-  for (const [x, y] of heartPositions) {
-    setCell(grid, x, y, 3, 1) // Heart character
-  }
-
-  return grid
-}
-
-/**
- * Simple test pattern for VIEW command testing
- * Numbers and letters to verify correct rendering
- */
-function createTestPatternGrid(): BgGridData {
-  const grid = createEmptyGrid()
-
-  // Row of numbers at top
-  for (let i = 0; i < 10; i++) {
-    setCell(grid, i + 9, 2, 48 + i, 2) // 0-9
-  }
-
-  // Row of letters
-  for (let i = 0; i < 10; i++) {
-    setCell(grid, i + 9, 4, 65 + i, 3) // A-J
-  }
-
-  // Color test - same character with different colors
-  for (let i = 0; i < 4; i++) {
-    setCell(grid, i + 12, 6, 35, i as 0 | 1 | 2 | 3) // # with 4 colors
-  }
-
-  // Simple smiley face
-  setCell(grid, 13, 10, 219, 2) // Left eye
-  setCell(grid, 15, 10, 219, 2) // Right eye
-  setCell(grid, 12, 12, 95, 2) // _
-  setCell(grid, 13, 12, 95, 2)
-  setCell(grid, 14, 12, 95, 2)
-  setCell(grid, 15, 12, 95, 2)
-  setCell(grid, 16, 12, 95, 2)
-
-  // Border
-  for (let x = 0; x < BG_GRID.COLS; x++) {
-    setCell(grid, x, 0, 45, 1) // -
-    setCell(grid, x, BG_GRID.ROWS - 1, 45, 1)
-  }
-
-  return grid
-}
-
-// ============================================================================
-// F-BASIC Reference Manual Sample Games BG Data (pages 94-101)
-// ============================================================================
-
-/** Empty cell marker for visual grids (3 underscores for visual alignment) */
-const ___ = '___'
-
-/** Letter to base CHR$ value mapping (from manual page 113) */
-const LETTER_BASE: Record<string, number> = {
-  A: 0, B: 8, C: 16, D: 24,
-  E: 184, F: 192, G: 200, H: 208,
-  I: 216, J: 224, K: 232, L: 240, M: 248,
-}
-
-// BG code constants for visual grids (from manual page 113)
-// Format: {letter}{digit}{color} where color is 0-3
-// K codes (232-239)
-const K72 = 'K72', K62 = 'K62', K52 = 'K52', K42 = 'K42'
-const K32 = 'K32', K22 = 'K22', K12 = 'K12', K02 = 'K02'
-const K50 = 'K50'
-// L codes (240-247)
-const L02 = 'L02', L12 = 'L12', L22 = 'L22'
-const L40 = 'L40'
-// J codes (224-231)
-const J00 = 'J00', J10 = 'J10', J20 = 'J20', J30 = 'J30'
-const J02 = 'J02', J12 = 'J12', J22 = 'J22', J32 = 'J32'
-const J50 = 'J50', J33 = 'J33', J03 = 'J03', J13 = 'J13', J23 = 'J23', J70 = 'J70'
-// I codes (216-223)
-const I60 = 'I60', I70 = 'I70', I62 = 'I62', I72 = 'I72'
-const I63 = 'I63', I73 = 'I73'
-// G codes (200-207)
-const G00 = 'G00', G10 = 'G10', G20 = 'G20', G30 = 'G30', G40 = 'G40'
-const G42 = 'G42', G52 = 'G52', G62 = 'G62'
-// D codes (24-31)
-const D41 = 'D41', D50 = 'D50', D52 = 'D52'
-// F codes (192-199)
-const F32 = 'F32', F41 = 'F41', F61 = 'F61', F72 = 'F72'
-// M codes (248-255)
-const M11 = 'M11', M31 = 'M31', M70 = 'M70', M71 = 'M71', M72 = 'M72'
-
-// ============================================================================
-// ASCII Character Constants for Visual Grids (3-char padded for alignment)
-// ============================================================================
-const _1_ = '1  ', _2_ = '2  ', _3_ = '3  ', _4_ = '4  ', _5_ = '5  '
-const _6_ = '6  ', _7_ = '7  ', _8_ = '8  '
-const _A_ = 'A  ', _B_ = 'B  ', _C_ = 'C  ', _D_ = 'D  ', _E_ = 'E  '
-const _F_ = 'F  ', _G_ = 'G  ', _H_ = 'H  ', _I_ = 'I  ', _J_ = 'J  '
-const _K_ = 'K  ', _L_ = 'L  ', _M_ = 'M  ', _N_ = 'N  ', _O_ = 'O  '
-const _P_ = 'P  ', _R_ = 'R  ', _S_ = 'S  ', _T_ = 'T  ', _U_ = 'U  '
-const _Y_ = 'Y  '
-
-/**
- * Convert a 3-char BG code from manual (e.g., "K72", "L02") to [charCode, colorPattern]
- */
-function parseBgCode(code: string): [number, 0 | 1 | 2 | 3] {
-  if (code.length !== 3) {
-    throw new Error(`Invalid BG code: ${code}`)
-  }
-
-  const letter = code[0]!
-  const digit = parseInt(code[1]!, 10)
-  const colorChar = code[2]!
-  const colorPattern = parseInt(colorChar, 10) as 0 | 1 | 2 | 3
-
-  const base = LETTER_BASE[letter]
-  if (base === undefined) {
-    throw new Error(`Unknown letter in BG code: ${code}`)
-  }
-
-  const charCode = base + digit
-  return [charCode, colorPattern]
-}
-
-/**
- * Create a cell from ASCII character
- */
-function asciiCell(char: string, color: 0 | 1 | 2 | 3 = 0): [number, 0 | 1 | 2 | 3] {
-  return [char.charCodeAt(0), color]
-}
-
-/**
- * Convert a visual 2D string grid to BgGridData
  *
- * Cell format:
- * - "___" (3 underscores) = empty cell
- * - "K72" = BG code (letter + digit + color pattern)
- * - "A  " = ASCII character (letter + 2 spaces)
- *
- * @param visualGrid - 2D array of 3-char cell strings
- * @returns BgGridData ready for use
+ * Grid data is split into two files by category:
+ * - sampleBgDataSimple.ts — programmatic grid creators (IDE demo samples)
+ * - sampleBgDataManual.ts — F-BASIC reference manual sample game grids (pages 94-101)
  */
-function visualGridToBgGridData(visualGrid: string[][]): BgGridData {
-  const result = createEmptyGrid()
 
-  for (let y = 0; y < visualGrid.length && y < BG_GRID.ROWS; y++) {
-    const row = visualGrid[y]
-    if (!row) continue
+import type { BgGridData } from '@/features/bg-editor/types'
 
-    for (let x = 0; x < row.length && x < BG_GRID.COLS; x++) {
-      const cell = row[x]
-      if (!cell || cell === ___) continue
-
-      const trimmed = cell.trim()
-      if (trimmed.length === 0) continue
-
-      // Check if it's a 3-char BG code (letter + digit + color) or ASCII text
-      if (/^[A-M][0-7][0-3]$/.test(trimmed)) {
-        // BG code like K72, L02, etc.
-        const [charCode, colorPattern] = parseBgCode(trimmed)
-        setCell(result, x, y, charCode, colorPattern)
-      } else {
-        // ASCII text - use first character
-        const [charCode, colorPattern] = asciiCell(trimmed[0]!, 0)
-        setCell(result, x, y, charCode, colorPattern)
-      }
-    }
-  }
-
-  return result
-}
-
-/**
- * KNIGHT game BG (page 94)
- * 8x8 chessboard with labels A-H and 8-1
- */
-function createKnightGrid(): BgGridData {
-  const grid: string[][] = [
-    //       0    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, _8_, ___, _7_, ___, _6_, ___, _5_, ___, _4_, ___, _3_, ___, _2_, ___, _1_, ___, ___, ___, I60, J30, J30, J30, J30, J30, J30, I70, ___],
-    /*  2 */ [K72, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, L02, ___, ___, J20, _K_, _N_, _I_, _G_, _H_, _T_, J20, ___],
-    /*  3 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _A_, ___, J00, J30, J30, J30, J30, J30, J30, J10, ___],
-    /*  4 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  5 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _B_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _C_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _D_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _E_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _F_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _G_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _H_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [L12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * SUPER MEMORY game BG (page 95)
- * Four L-shaped corner panels pointing toward center
- */
-function createSuperMemoryGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, I60, J30, J30, J30, J30, J30, I70],
-    /*  2 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _S_, _U_, _P_, _E_, _R_, J50],
-    /*  3 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _M_, _E_, _M_, _O_, _R_, _Y_],
-    /*  4 */ [___, K52, K52, K52, L22, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, L12, K52, K52, K52, ___, J00, J30, J30, J30, J30, J30, J70],
-    /*  5 */ [___, K52, K52, K52, K52, L22, K62, ___, ___, ___, ___, ___, ___, ___, K62, L12, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [___, K52, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, K62, ___, ___, ___, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [___, K52, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [___, K52, K52, K52, K52, L02, K62, ___, ___, ___, ___, ___, ___, ___, K62, K72, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [___, K52, K52, K52, L02, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K72, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * UFO game BG (page 96)
- * Starry sky with mountain terrain at bottom
- */
-function createUfoGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, ___, ___, M11, M31, ___, ___, ___, ___, G62, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  2 */ [___, F41, D41, D41, D41, D41, F61, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, G52, ___, ___, ___],
-    /*  3 */ [___, ___, K12, K52, K52, K12, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  4 */ [___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, G62],
-    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, G62, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___],
-    /*  9 */ [___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G52, G62, ___, ___, ___, ___, ___, G62, ___, G52, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___],
-    /* 12 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___],
-    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, G00, G20, G30, G10, ___, ___, ___],
-    /* 16 */ [___, ___, ___, ___, ___, ___, ___, G00, G20, G30, G10, ___, ___, ___, ___, ___, ___, ___, G00, G20, G20, G20, G20, G30, G30, G10, ___, ___],
-    /* 17 */ [G40, M70, M70, G40, G40, M70, G20, G20, G20, G40, G30, G40, G40, M70, G40, M70, G40, G40, G40, G40, G20, G20, G40, G40, G30, G30, G40, G40],
-    /* 18 */ [M70, G40, M70, M70, G40, G40, G40, G40, M70, M70, G40, G40, M70, G40, M70, M70, M70, G40, M70, M70, G40, G40, M70, G40, G40, G40, M70, M70],
-    /* 19 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-    /* 20 */ [M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * ROUTE 66 game BG (page 97)
- * Road with edges - all rows same pattern
- */
-function createRoute66Grid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  1 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  2 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  3 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  4 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  6 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 16 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 17 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * TYPE MASTER game BG (page 98)
- * Text boxes and frames for typing display
- */
-function createTypeMasterGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, K72, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, K62, _T_, _Y_, _P_, _E_, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  2 */ [___, L12, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  3 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, ___, ___, ___, ___],
-    /*  4 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
-    /*  5 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
-    /*  6 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, I62, J32, I72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, J22, ___, J22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, J02, J32, J12, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, I63, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, I73, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [___, J03, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J13, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * TURTLE game BG (page 99)
- * Racing track with lanes and finish line
- */
-function createTurtleGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50],
-    /*  1 */ [D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52],
-    /*  2 */ [G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42],
-    /*  3 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
-    /*  4 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
-    /*  5 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K22],
-    /*  6 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /*  7 */ [_1_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /*  8 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /*  9 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 10 */ [_2_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 11 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 12 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 13 */ [ _3_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 14 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 15 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 16 */ [ _4_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 17 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 18 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 19 */ [ _5_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 20 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K12],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * CARD game BG (page 100)
- * 6x6 card grid frame
- */
-function createCardGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, ___, ___, ___, ___],
-    /*  1 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  2 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  3 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /*  4 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  5 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  6 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /*  7 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  8 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  9 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
-    /* 10 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 11 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 12 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
-    /* 13 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 14 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 15 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /* 16 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 17 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 18 */ [L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * SCR$ Sample game BG (page 101)
- * Maze with flags (F72) and brick walls (F32)
- */
-function createScrSampleGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-    /*  1 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32],
-    /*  2 */ [F32, ___, ___, F72, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /*  3 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  4 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, F32],
-    /*  5 */ [F32, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, F32, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /*  6 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, F72, ___, ___, F32, F32, ___, F72, ___, F32],
-    /*  7 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  8 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  9 */ [F32, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 10 */ [F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /* 11 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 12 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 13 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, F72, ___, F32],
-    /* 14 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, F72, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 15 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /* 16 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /* 17 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 18 */ [F32, ___, ___, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /* 19 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 20 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-// ============================================================================
-// Export Sample BG Data Map
-// ============================================================================
-
-/**
- * Layer visualization box frame - simple border to show BG GRAPHIC boundaries
- * Used with printableArea sample to visualize all 3 screen layers
- */
-function createLayerBoxGrid(): BgGridData {
-  const grid = createEmptyGrid()
-
-  // Draw a simple box frame using CHR$(254) - different from CHR$(255) used in printable area
-  // Top and bottom borders
-  for (let x = 0; x < BG_GRID.COLS; x++) {
-    setCell(grid, x, 0, 254, 1) // Top border
-    setCell(grid, x, BG_GRID.ROWS - 1, 254, 1) // Bottom border
-  }
-
-  // Left and right borders
-  for (let y = 0; y < BG_GRID.ROWS; y++) {
-    setCell(grid, 0, y, 254, 1) // Left border
-    setCell(grid, BG_GRID.COLS - 1, y, 254, 1) // Right border
-  }
-
-  return grid
-}
+import {
+  createCardGrid,
+  createKnightGrid,
+  createRoute66Grid,
+  createScrSampleGrid,
+  createSuperMemoryGrid,
+  createTurtleGrid,
+  createTypeMasterGrid,
+  createUfoGrid,
+} from './sampleBgDataManual'
+import {
+  createDemoBorderGrid,
+  createLayerBoxGrid,
+  createPlatformLevelGrid,
+  createTestPatternGrid,
+  createTitleScreenGrid,
+} from './sampleBgDataSimple'
+import { createEmptyGrid } from './sampleBgDataSimple'
 
 /**
  * Map of sample keys to their pre-designed BG data
